This approach might work better.
